requirements.txt
apscheduler==3.10.4
supabase==2.15.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.128.0
google-ai-generativelanguage
google-api-core
google-auth==2.47.0
google-genai==1.57.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
gunicorn
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
nba_api==1.11.3
numpy==2.4.1
pandas==2.3.3
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uriTemplate==4.2.0
urllib3==2.5.0
uvicorn
websockets==15.0.1
